weather_stats_API/requirements.txt
Flask
flask-restx
snowflake-connector-python
flask-restful
Flask-SQLAlchemy
flasgger